(WIP) Implement --dev mode for shell hooks in test_activate
Without this we end up running a mixture of shell scripts from the
dev sources and shell scripts from the conda installed in sys.prefix
with a conda.exe from sys.prefix (potentially without the dev sources
installed), leading to an illogical, undebuggable, undevelopable mess.

Instead I have gone for the same approach as the other --dev modes I have
added, namely, CONDA_EXE is either conda or Python. When it is Python
we also set _CE_M and _CE_CONDA to "-m" and "conda" respectively. When
we invoke CONDA_EXE it is *always* as:

"${CONDA_EXE}" ${_CE_M} ${_CE_CONDA}

.. or the Windows equivalent.

Please think about what code you are running and what rules you are
enforcing (even if not deliberate), please remember that explicit is
better than implicit and that wasting developer's time like this is
a terrible thing to do.

You might say "No Ray, you are wrong, you are only supposed to run
conda tests in an env that already has conda installed in develop
mode" and while that is an OK restriction, the testsuite should refuse
to run if that pre-condition is not met. But more importantly, given
that the shell/batch scripts are located relatively to CONDA_EXE it
makes for a terrible workflow on Windows when trying to fix those
scripts.

KNOW WHAT YOU ARE RUNNING, RUN ONLY THE THING YOU ARE DEVELOPING. THAT IS ALL.