Fix test_bash_activate_error for shells that already activate conda
We can make both cases succeed so that is what we do. The issue is that mounting
of /c on MSYS2 is handled by login scripts so we must pass -l to bash. However if
a user had conda init stuff in ~/.bash_profile then their shell will start activated