Add upgrade flag to pip installer
The "-U" flag ensures that pip-managed packages are always upgraded (within the requirements) when "conda env update" is used.

Adding the flag here causes it to always be used, even for "conda env create", but in that case the "-U" has not effect because the package is not already installed.