allow conda to be downgraded with explicit spec
Allow conda to be downgraded if by the user with an explicit spec.
For example conda 4.7.6 can be downgraded to 4.6.14 if a user runs:
    conda install conda=4.6.14
Since ea6aae6c21d04b057404cfc8cea6efe50bcc574e in #8741
conda was only downgraded with an explicit spec if conda was configured
not to automatically update.

closes #8824