target installed python package
When adding in logic to prevent python from being updated beyond the
current minor version include the currently installed python package
as a MatchSpec target. This instructs the solver not to disturb this
package unless necessary. Without this change python is a requested
package whose version is maximized within the bounds early in the solve
process.