add prefix pkg to index if channels do not match
Add the local package to the index with the a canonical channel set to
the URL if the channel information of the local packages does not match
that of the remote package.

Fixes cases where each channel in a multi-channel has a package with the
same subdir, name, version and build_number.

closes #8302