Add a couple of unit tests for caching of metadata
The first one confirms that if we create SubdirData object for
a channel and then do that again, the second object is actually
the same one as the first; it's been fetched from the in-memory cache.

The second test confirms that if we do the same thing but clear the
cache between the creation of the first and second objects then we get
two distinct objects, but they're equivalent; a query on each of them
yields the same result.