fixup the documentation for the --channel option
This improves the documentation for the `--channel` argument to ``conda
build`. The `file://` syntax is not the only way to specify local
directories. It turns out that you can simply list an absolute or
relative path. This can come in handy when building with a Unix shell,
like git-bash, on Windows where drive letter and forward-/back-slash
conventions might lead to confusion due to the insertion of spurious
characters and/or protective escapes.