Discard cached metadata if JSON file is newer
We now timestamp each cached metadata object with the time it was placed
in the cache.  When we're about to use a cached object from a
"file:///" channel, we first check if the JSON file from which it was
loaded still exists, and get its modification time.   If the JSON file
no longer exists, or if the JSON file is newer than the cached object,
then we ignore the cached object and create a new one (which will
replace the old one in the cache).