logz: Swap pformat to json.dumps when stringifying json response logs
When wishing to (debug) print 256 characters from repodata.json from conda-forge,
pprint() was observed to take 70 seconds. Changing this to json.dumps() gets it
to taking less than a second. Adding indent=2 brings that back up again to a
couple of seconds, so to counteract that, we make the assumption that every dict
item will result in at least one character being emitted. This seems reasonable
to me.