Propose clarifcation on merge strategy.
This section is a big vague. Could you clarify this section a bit more?

I am not sure if my understanding is correct:
My best guess is that for a list key 'merge' here means extend, and for a dict 'merge' means update. (Per python method names).

It however could also be possible merge here means recursively merge items of a list / values of a dict.