Support nested keys in conda config --get
The bug:
$ echo "conda_build:">>~/.condarc
$ echo "  cache_dir: /opt/conda/conda-bld">>~/.condarc
$ conda config --get conda_build.cache_dir
unknown key conda_build.cache_dir

The diagnosis:
We use nested keys in conda-build, but conda refuses to play ball
because it cares not for our nesting when it looks for the key.

The fix:
Step through the requested, dotted, nested key and rc_config in
lock-step.

The resolution:
$ conda config --get conda_build.cache_dir
--set conda_build.cache_dir /opt/conda/conda-bld