Fix changeps1 implementation for Powershell
Since commit ca16da40 using changeps1=false just results in
Add-CondaEnvironmentToPrompt not being called, but still unconditionally
renaming the global prompt function so the end result is the user
ends up with no prompt at all (well, just 'PS>').
Fix this by making everything related to the prompt conditional, but
execute all of that in the module itself (and add arguments to the
module in order to achieve that); see previous issues related to the
prompt function: it should be altered in the global scope and not
in function scope.