Refactoring tests
This commit introduces some changes based on testing. From certain
failures that were previously in the tests, it makes sense to disable
the notices decorator when `offline` is set to true.

The other changes involve removing the `conftest.py` files and instead
putting this code in its own module under `conda.testing.notices`. These
are then added either by importing them, or for the case of fixtures,
including them as a pytest plugin in the root `conftest.py` file in our
tests.