Moving out of env, updating tests
This commit moves the rename command from conda_env to conda. It also
modifies how we test the rename command. I have decided to convert these
to pytest tests and I have also chosen to use the subprocess module for
running the commands because it was a simpler implementation than using
the available helper utilities under "conda.testing.integration".