Update Dockerfile + ci-images flow to build multi arch images
This is done to have linux/aarch64 conda-ci images, which perform better
on Apple M1 machines compared to the linux/amd64 images.

The other images (linux/ppc64le, linux/s390x) we basically get for free
and prove that the whole tooling is consistently available on all 4 linux
platforms.

Once the multi-arch image is published it can be also used during the ci
flow to e.g. run a short smoke test on all the non-native platforms via
emulation.