Fix GIT_DESCRIBE_REGEX to strip non-numeric characters at beginning.
This regex is used to parse the output of `git describe` and turn it
into a version, post commit number, and commit hash that can be
formatted into a PEP440-compliant version number. The first part of this
regex was presumably intended to match letters, '_', or '-' at the
beginning of the string to strip them off exclusive of the version
number. The pattern was incorrect, however, with an unescaped '-' being
interpreted as a character range instead of the literal dash character.
The result was that all lowercase letters except 'z' would not be
stripped from the version.

The common case that this fixes is a git version tag that has a 'v'
prefix, like "v1.0.0". Previously, this would result in an invalid
version number of "v1.0.0". With this fix, the version number becomes
"1.0.0".