CEP-6 Channel Notices (#11462)
* saving work so far on this new feature

* saving progress

* adding more work that gets this PR to a workable state

* fixing import errors

* switch to the dals function for string formatting help

* adding documentation for the new command

* Re-organizes the code

I felt that the notices.py file was getting a little too large. To get
around that, I implemented a new organizational structure which includes
a new top level "apps" module. This follows more-or-less what Django
does with its structure, and I think this pattern could help the conda
repository too.

* Refactoring test code

This commit refactors the test code to be a little more DRY. It also
establishes a single place for helper functions and pytest fixtures for
testing the notices code.

* Adding user story tests

This commit adds tests but in a new "user story" format I was playing
around with.

* Refactors and testing

This pull request removes the SQLite db approach to caching seen ids and
instead relies on a text file. It also add more testing. At this point
the test coverage is at 100% for the notices module.

* addressing issues with tests

* Refactoring tests

This commit introduces some changes based on testing. From certain
failures that were previously in the tests, it makes sense to disable
the notices decorator when `offline` is set to true.

The other changes involve removing the `conftest.py` files and instead
putting this code in its own module under `conda.testing.notices`. These
are then added either by importing them, or for the case of fixtures,
including them as a pytest plugin in the root `conftest.py` file in our
tests.

* Changes configuration parameters

This commit removes the `disable_channel_notices` configuration
parameter and replaces it with the `number_channel_notices`. The reason
behind this is so we could control how many messages are shown and
completely disable notices with a single parameter.

* removing mentions of interval as it is an unused field

* removing use of TypeDict

* resolves issue with windowz cache dir

* updating tests that break with windowz

* removing file

* fixing import of mock

* Updating to use "expired_at" instead of "expiry"

This commit switches the behavior of the expiry for notices. Instead of
defining the "expiry" field as an integer representing seconds, we
instead directly use a timestamp field.

* Apply suggestions from code review

Co-authored-by: Jannis Leidel <jannis@leidel.info>

* More changes from recent code review

Mostly just fixing imports to use relative imports and also a couple of
other structural things with the code.

* Fixing tests to improve coverage

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* implementing better fuzzy match for mock http logger calls

* ensuring notices will not show when --json flag has been passed

* Adding in missing mock notices cache file for Windows

Hopefully this will address remaining testing issues for Windows in the
CI.

Co-authored-by: Jannis Leidel <jannis@leidel.info>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>