Add safer behavior when removing destination directories
Before this commit, if there was an error during the rename process
while the --force flag was passed, the destination directory would be
removed and not recovered. This commit implements a safer way of doing
this that has the ability to recover from error and return the
filesystem to the way it was before running the command.