Updates my previous suggestions for improvements
This time we have an implementation that fully supports any CLI parsing
library you may want to use. It does by modifying the ArgumentParser
class to add a "pass-thru" for plugin subcommands. This simply means
that we add a stub for the command and then pass it off to the plugin
for actual command line argument parsing.