Require `toolz` & soften `cytoolz` dependency
As the codebase can benefit from `cytoolz` when present and still
fallback to `toolz` when `cytoolz` is absent, would suggest softening
the dependency here to just require `toolz`. This can be particularly
helpful when supporting other Python implementations (like PyPy, etc.)
since `toolz` is pure Python and should just work out-of-the-box. Thus
this simplifies the process of bootstrapping a new Python implementation
with Conda. Additionally if one or the other is faster for a Python
implementation, this leaves it up to users to decide.