Resolve prefix injection (#11666)
Removed prefix injection logic originally introduced in conda 4.4 to compensate for a Python 3.6 packaging patch on Windows (this patch has long since been removed and was not present in other Python versions). This original logic evolved over time to be a generic prefix injection that occurred on the shell level and was applied on all OSs. Ultimately the prefix injection resulted in inconsistent behaviors between `conda activate ENV ; <command>` and `conda run -n ENV <command>`. In resolving this issue, we also resolve incorrect behavior in environment stacking.