Prototype for a conda session manager and context extension
This is a prototype for a conda session manager and how we would be able
to extend this using our plugin system. Additionally, it also defines a
plugin hook for extending the context object.

These implementations are intentionally incomplete as they are more
geared towards starting a conversation about the best way to implement
such things in the current conda codebase.