Added additional checks to determine if conda.csh set the prompt variable, and clear the setting when the script ends if set.
Description
The conda.csh activation script will set the csh prompt variable if not already set. The prompt variable is needed for the activate, deactivate, etc. subcommand. When run in non-interactive shells the current conda.csh script will leave the prompt variable set, which can cause issues for other scripts that run different commands based if the shell is interactive or not.

This fix moves the setting of prompt outside the if/else block, and then unsets the prompt variable if the script did set it.

Fixes #9281