Fixes issue with tests plus more
The order of decorator functions was causing an issue with tests. Before
this commit we had the following order:

cache -> env_var_override

Now it is:

env_var_override -> cache

This means that the env_var_override will always override the value
where before this was being cached.

Also changed in this commit is the adding back of the `cuda_version` to
the `context` object. I figured this would be a good idea because
removing this may break other projects out there that have come to
depend on it.