Remove broken logic for setting `for_user`
Looking back at the commit history, I think:
```
        if not (args.install and args.user and args.system):
            for_user = True
```
used to make sense when `--no-user` didn't exist and `--user` was
not the default (it had to be specified).  Even then, I think the
logic was probably wrong.  I think it was probably meant to say,
if none of `--user`, `--install` and `--system` is provided, do a
user init.  But it actually said if all 3 flags are not provided,
do a user init.  This is odd because `--user` and `--system`
don't seem to make sense at the same time.

My sense at this point is that this line just doesn't make sense
anymore at all and should be removed.