Adds minimal documentation and test for misconfiguration
This commit adds the minimal amount of documentation I feel is necessary
for this new feature. I don't want to add too much for now as I feel it
will be better to give this a "silent" release initially.

Additionally, I've edited the "PrimitiveParameter.load" method to do
some checking for which types it loads. This will provide better error
handling when people misconfigure not only channels but other
configuration parameters too.