Fixes slow running test
For this test, it's not actually necessary for us to even run the
install command. We just want to check whether the notices are running
as we expect them to.

So, instead of relying on the fixture that runs conda-build under the
covers, we've removed it and instead just mock out the install function.