Fix behavior of `--no-user` flag in `conda init` (#11949)
* Fix behavior of `--no-user` flag in `conda init`

It is clear from the code that `args.no_user` should have a value
of `True` when the `--no-user` flag is provided, so that the
user's `.bashrc`, etc. is not updated.

* Switch to `--no-user` sets `args.user` to `False`

This is more intuitive to follow through the code but has the
complication that the default mush be handled differently.

* Set `default=True` for `--user`

* Remove broken logic for setting `for_user`

Looking back at the commit history, I think:
```
        if not (args.install and args.user and args.system):
            for_user = True
```
used to make sense when `--no-user` didn't exist and `--user` was
not the default (it had to be specified).  Even then, I think the
logic was probably wrong.  I think it was probably meant to say,
if none of `--user`, `--install` and `--system` is provided, do a
user init.  But it actually said if all 3 flags are not provided,
do a user init.  This is odd because `--user` and `--system`
don't seem to make sense at the same time.

My sense at this point is that this line just doesn't make sense
anymore at all and should be removed.

* If `--system` is specified disable user init