Speedup pre-commit by factor 15
This removes pylint and bandit from pre-commit, as both are slow and their
feedback is anyway already ignored, due to running with exit-zero.

This locally reduces the pre-commit runtime from ~43sec to 2.9sec and thus
makes it possible to run pre-commit in a loop during local development to
constantly provide feedback / style the code.