Implement 2-phase parser with greedy subparsers (#12910)
Implement a 2-phase parser so we can properly intercept `--no-plugins` before loading the subcommand plugins. In addition to the 2-phase parser we add a new subparser that greedily consumes all unknown arguments to be passed to the subcommand plugins. These changes allows a simplification of how third-party subcommands are detected and displayed in the help text (whether subcommand plugins or legacy executables) so all subcommands are now processed via conda's argparse parser instead of hijacking argparse's error handling.

---------

Co-authored-by: Jannis Leidel <jannis@leidel.info>