Convert canary builds to be noarch
There is no technical reason why a package build on macos-{64, arm64} or
linux-{64, ppc64le, s390x} should be any different. So building two noarch
variants of conda, one for unix and one for windows. The run constrain
ensures that only the right variant can be installed.

While its is the goal of prod/defaults to build recipes for all platform /
python combination to e.g. reveal if dependencies are missing in some of
the combinations, the the goal of conda canary releases may differ in this
regards to make canary builds as easy testable as possible on all platforms
/ python combinations.