Add sdist to release uploads
#13663 added the workflow to upload a reproducible release tarball to the
release artifacts to not rely on the non-stable Github mechanisms that
provide those.

While archiving the entire repo via `git archive` is nice, the real
Pythonic artifact is a sdist (Source Distribution) file as also
packages upload to pypi.

This PR adds the sdist creation to the upload flow and also changes the
sha256sum calculation to produce a combined SHA256SUMS files, as the
sdist file name is not something we influence and can rely on in the flow,
but is given by hatch and the configuration in the pyproject.toml. Sotagged
releases produce a conda-$tag.tar.gz (size 550 KB, name produced by hatch) and
conda-$tag_full.tar.gz (size 21 MB, name defined in the flow) file and a combined
SHA256SUMS file (huge size difference!).