import all modules before env update is run
In the 'conda env update' flow, import all modules that might be used
in the future before updating the environment. In this way the modules
will be present in the cache so that they can be used.  Since an update
can change conda itself importing from the filesystem may not be
possible.