Remove empty PATH entry on Windows
With the Windows 10 release, Microsoft added an interactive editor for the PATH variable e.g. [the workflow described here](https://learn.microsoft.com/en-us/dotnet/core/install/windows?tabs=net80#no-net-sdk-was-found). This normalizes the PATH entry by appending a `;` at the end of all entries. Here, because we split on the `;`, we'll end up with an empty entry. Later in the activation code, the paths are normalized with `os.path.normpath` which will expand this to `.` which can affect downstream code and can cause users concern since `.` has a special meaning on POSIX based systems.