Windows 10+ relevance
Windows 10+ operating systems no longer have a "ProgramData" folder on the C:\ drive, so that is an obsolete reference. Also, miniconda (the free distribution of Anaconda, no license fees required and thus the most popular moving forward indefinitely) always installs for the local user in the "AppData/Local" subfolder of the user's profile.