Add `conda doctor` check for if the pinned file is ill formatted (#15022)
* Refactor get_pinned_specs to be part of prefix data

* Add health check for conda-meta/pinned file

In this check, we'll inspect each pinned package. If the package
does not exist in the prefix, that might indicate that the package
is not written out right. For example, this may catch typos or
extra punctuation added to package names.

* Add news

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add constant for path to pinned file in prefix

* Update conda/plugins/subcommands/doctor/health_checks.py

Co-authored-by: Jannis Leidel <jannis@leidel.info>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jannis Leidel <jannis@leidel.info>