Bump conda-libmamba-solver from 24.11.0 to 25.4.0
In Conda 25.9.0, some deprecated functionality that earlier versions of
conda-libmamba-solver used was removed. While conda-libmamba-solver
updated its logic and released 25.4.0, Conda kept the older lower bound
of 24.11. As a result users could update Conda, but still get the older
conda-libmamba-solver. This would lead to issues running Conda after. To
fix this, bump the lower bounds of conda-libmamba-solver in Conda.